taskModel

+++ src/server/project.js
@@ -1,9 +1,10 @@
 import {
     projectDB,
+    taskDB,
 } from './index';
 
 export function getProjectList (params) {
-    let query = {};
+    const query = {};
     if (params.status !== -1) {
         query.status = params.status;
     }
@@ -28,18 +29,53 @@ export function doCreateProject (data) {
     });
 }
 
-export function doDeleteProject (_id) {
+export function getProjectInfo (_id) {
     return new Promise((resolve, reject) => {
-        // TODO 没有task的才允许删除
-
-        projectDB.remove({
+        projectDB.findOne({
             _id,
-        }, {}, (err) => {
+        }, (err, doc) => {
             if (err) {
-                console.error(err);
+                reject(err);
+            }
+            resolve(doc);
+        });
+    });
+}
+
+export function doEditProject (data) {
+    return new Promise((resolve, reject) => {
+        projectDB.update({
+            _id: data._id,
+        }, data, {}, (err) => {
+            if (err) {
                 reject(err);
             }
             resolve();
+        });
+    });
+}
+
+export function doDeleteProject (_id) {
+    return new Promise((resolve, reject) => {
+        taskDB.count({
+            projectId: _id,
+        }, (err, count) => {
+            if (err) {
+                console.error(err);
+                return reject(err);
+            }
+            if (count > 0) {
+                return reject(count);
+            }
+            projectDB.remove({
+                _id,
+            }, {}, (err) => {
+                if (err) {
+                    console.error(err);
+                    reject(err);
+                }
+                resolve();
+            });
         });
     });
 }
